Started the CSV Reader and Base Card Work In Progress
Need to hammer out the image issues and make it run better overall. This is currently the first pass. CSV reader will spawn instantiated cards, but need to make the image processing faster and more accurate. Created the base card class as well, will probably make child classes of it depending on the types of cards we need.