Added a lot of logic in for the refactor and Cards, and CSV Reader
Pushing in case I need to work from home on Monday. It is currently a work in progress. Need to finish the ECS of the cards, make the csv better and in built in directory, and add in more card images and text. Card play is working temporarily, but need to fix discard.