Moved the malicious actor. And implemented a lot more of the card play for both of the malicious actor and the resilient players
Moved the malicious actor off of the game manager to make sure they can have their child cards parented and located properly. Added duration to the cards to make sure they despawn at the proper point. Need to incorporate their actual potency/impact. Also need to change action costs and the continuation of the building of the ECS of the cards and resolve 2 errors that occur on spawn that don't impact the game. Also updated slippy to allow for the cards to be moved properly and not impact the regular slippy version of the map.