Split up resilient and malicious base cards, fixed card play for malicious actor,
Split up the base card to have 1 base card for both teams to fit their designs. The implementation and design of their card is still a work in progress. Also converted the cardfront struct to a class. Also did a lot of development to sophisticate the card play, especially for the malicious actor by letting them select more than 1 facility and also not be able to play cards on a facility they targetted that turn. Still need to do this for the resilient player and also remove the targetIDs after their duration expires.