Larger changes to card play and facility distribution
Now players will not have any shared access to facilities, as they will be distributed to their own player. Made some changes to the card design and also incorporating some of the newer aspects of the csv into the card reader. Also enhanced how the cards are played to make it more accurate in how it responds and reacts. Need to sophisticate it more to handle the differences in card types. Also added in enums for the diff card types to allow for when being read in to handle their impacts better.