Sophisticated Card play a bit, added in hand locations, ensured duration is working
Developed the card play to interact with each other better. Implemented and got the hand locations working, so that the cards will be magnetized to that standard spot, unless if they are placed on the card drop location and have the right number of targets selected. With both of these working, I was able to test to ensure the implemented duration mechanics worked properly.