improved card readability and a card moving bug
Card description has been increased in size and offset to appear above the card instead of overlapping with the card itself. this has only been done to the malicious cards so far so we can test to see which one we like more

There was a bug where if you move the cards around in your hand they would shrink or grow in size. this has been fixed and they will be set to their original size when dropped. its still a bit off but much better